ml/requirements.txt
# ML surplus prediction pipeline
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
# Optional: for loading from vector DB
pinecone-client>=3.0.0
